square works

+++ app/controllers/application_controller.rb
@@ -7,7 +7,7 @@ class ApplicationController < ActionController::Base
 
   def calculate_square
 
-    @num = params.fetch("elephant").to_i
+    @num = params.fetch("elephant").to_f
     @square_of_num = @num ** 2
 
     render({ :template => "calculation_templates/square_results_form.html.erb"})
